+++ src/args/mod.rs
@@ -2,3 +2,4 @@ pub mod args;
 pub mod default_subcommand;
 pub mod prefs_subcommand;
 pub mod cmd;
+pub mod settings_subcommand;