+++ Changelog.md
@@ -16,3 +16,9 @@
 
 ## 0.3.2
 - Don't panic when the update check fails due to network issues
+
+## 0.4.0
+- Support rewriting launch properties, prefs to set default ui scale `prefs set scale 2`
+
+## 0.5.0
+- Experimental unix-only support for backing up and restoring Ghidra preferences